tdd: support with VS202x
Add `zephyr_gtest_uart_stm32`, skip function call to
`arch_tls_stack_setup` in VS202x.

+++ lib/net_buf/buf.c
@@ -79,7 +79,7 @@ static inline struct net_buf *pool_get_uninit(struct net_buf_pool *pool,
 
 	buf = (struct net_buf *)(((uint8_t *)pool->__bufs) + byte_offset);
 
-	buf->pool_id = pool_id(pool);
+	buf->pool_id = (uint8_t)pool_id(pool);
 	buf->user_data_size = pool->user_data_size;
 
 	return buf;
